requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.0
black==22.6.0
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
distlib==0.3.5
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
filelock==3.8.0
flake8==5.0.4
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
identify==2.5.3
idna==3.3
interrogate==1.5.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
orjson==3.8.0
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.1
pydantic==1.9.2
pyflakes==2.5.0
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
tabulate==0.8.10
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.16.3
watchfiles==0.16.1
watchgod==0.8.2
websockets==10.3
